Separate language service interfaces into their own file
Currently, `ast/analyzer.ts` is a hodge-podge of interfaces that are
useful when implementing a Jsonnet language service.

This commit will continue down the road of separating individual
components out into their constituent parts. Specifically, in this
commit we will move interfaces related to the language service, i.e.,
the code that consumes events emitted when a user interacts with the
development environment, and emits things like autocomplete suggestions.
For example: `UiEventListener`, which keeps hooks when something like a
`Hover` event is fired, or `CompletionInfo`, which represents an
autocomplete suggestion.
